requirements.txt
Django==1.8.3
PyJWT==1.4.0
PySAL==1.11.0
chardet==2.2.1
colorama==0.2.5
decorator==3.4.0
django-cors-headers==1.1.0
django-filter==0.12.0
django-pgjson==0.3.1
djangorestframework==3.3.2
djangorestframework-gis==0.10.1
djangorestframework-jwt==1.7.2
html5lib==0.999
httplib2==0.9.2
numpy==1.10.4
psycopg2==2.4.5
requests==2.9.1
scipy==0.13.3
six==1.5.2
urllib3==1.7.1
wheel==0.24.0